+++ tests/__init__.py
@@ -1,0 +1,7 @@
+from .default import *
+from .singlefiles import *
+from .strangenames import *
+
+from validator.utils import setup_logger
+
+setup_logger(True)