+++ tests/default.py
@@ -17,15 +17,15 @@ class TestDefault(TestCase):
         os.chdir(cls.old_cwd)
 
     def test_doubleQuotation(self):
-        (status, successful, failed, ignored) = validator.validate('"failing/failing-validation.version"')
+        (status, successful, failed, ignored) = validator.validate_cwd('"failing/failing-validation.version"')
         self.assertEqual(status, 0)
 
     def test_invalidWorkspace_recursive(self):
-        (status, successful, failed, ignored) = validator.validate('')
+        (status, successful, failed, ignored) = validator.validate_cwd('')
         self.assertEqual(status, 1)
 
     def test_exclusionWildcard(self):
-        (status, successful, failed, ignored) = validator.validate('failing/*.version')
+        (status, successful, failed, ignored) = validator.validate_cwd('failing/*.version')
         self.assertEqual(status, 0)
         self.assertSetEqual(successful, {Path('default.version'), Path('recursiveness/recursive.version'),
                                          Path('recursiveness/recursiveness2/recursive2.version')})
@@ -33,7 +33,7 @@ class TestDefault(TestCase):
         self.assertEqual(failed, set())
 
     def test_excludeAll(self):
-        (status, successful, failed, ignored) = validator.validate('["./**/*"]')
+        (status, successful, failed, ignored) = validator.validate_cwd('["./**/*"]')
         self.assertEqual(status, 0)
         self.assertSetEqual(successful, set())
         self.assertSetEqual(ignored, {Path('default.version'),
@@ -43,7 +43,7 @@ class TestDefault(TestCase):
         self.assertEqual(failed, set())
 
     def test_recursiveExclusion(self):
-        (status, successful, failed, ignored) = validator.validate('["./recursiveness/**/*"]')
+        (status, successful, failed, ignored) = validator.validate_cwd('["./recursiveness/**/*"]')
         self.assertEqual(status, 1)
         self.assertSetEqual(successful, {Path('default.version')})
         self.assertSetEqual(ignored, {Path('recursiveness/recursive.version'),
@@ -51,7 +51,7 @@ class TestDefault(TestCase):
         self.assertEqual(failed, {Path('failing/failing-validation.version')})
 
     def test_multipleExclusions(self):
-        (status, successful, failed, ignored) = validator.validate('["./*.version", "./failing/*"]')
+        (status, successful, failed, ignored) = validator.validate_cwd('["./*.version", "./failing/*"]')
         self.assertEqual(status, 0)
         self.assertSetEqual(successful, {Path('recursiveness/recursive.version'),
                                          Path('recursiveness/recursiveness2/recursive2.version')})