+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (not included in any release yet)
 * Validate remote version file if specified with `URL` property.
+* Make the validator importable as a package (#4 by: DasSkelett)
+* Use logging instead of print(), make use of the Action logging syntax (#4 by: DasSkelett)
 
 
 ## v1