+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Make the validator importable as a package (#4 by: DasSkelett)
 * Use logging instead of print(), make use of the Action logging syntax (#4 by: DasSkelett)
 * Implement KSP version comparison logic + warn for outdated KSP compatibilities (#5 by: DasSkelett)
+* Throw validation errors only once; format 'requests' debug logging (#6 by: DasSkelett)
 
 
 ## v1