+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 # This file contains all notable changes to the AVC-VersionFileValidator
 
 ## master (not included in any release yet)
+
+
+## v1
+### v1.2.0
 * Validate remote version file if specified with `URL` property.
 * Make the validator importable as a package (#4 by: DasSkelett)
 * Use logging instead of print(), make use of the Action logging syntax (#4 by: DasSkelett)
 * Implement KSP version comparison logic + warn for outdated KSP compatibilities (#5 by: DasSkelett)
 * Throw validation errors only once; format 'requests' debug logging (#6 by: DasSkelett)
 
-
-## v1
 ### v1.1.1
 * Add requirements.txt for easier dev env setup.
 * Add examples in the examples/ folder. The standard.yml should cover most use cases.